Use size_t constants for joint count arithmetic
Agent-Logs-Url: https://github.com/3218158258h/glove_ros2/sessions/49417f1d-a2b6-44a5-b145-d7a3baca115d

+++ examples/cyclonedds/glove_dds_publisher.cpp
@@ -24,9 +24,9 @@ using namespace SGCore::Kinematics;
 
 static std::atomic<bool> g_stop{false};
 static constexpr std::chrono::milliseconds kPublishInterval{15}; // ~66.7 Hz
-static constexpr uint8_t kJointsPerFinger = 3;
-static constexpr uint8_t kFingerCount = 5;
-static constexpr uint8_t kMaxJointCount = kJointsPerFinger * kFingerCount;
+static constexpr size_t kJointsPerFinger = 3;
+static constexpr size_t kFingerCount = 5;
+static constexpr size_t kMaxJointCount = kJointsPerFinger * kFingerCount;
 static constexpr float kDefaultEulerValue = 0.0f;
 
 enum JointCode : uint8_t
